+++ src/Hooks/GameLevelManager.hpp
@@ -1,0 +1,7 @@
+#include <Geode/modify/GameLevelManager.hpp>
+
+using namespace geode::prelude;
+
+class $modify(ModGameLevelManager, GameLevelManager) {
+	void deleteLevel(GJGameLevel* level);
+};