+++ changelog.md
@@ -1,6 +1,7 @@
 ## Unreleased (1.4)
 - Added a Save Manager, hapi \[[Issue #5](https://github.com/Kevadroz/PracticeCheckpointPermanence/issues/5)\]
 - Disabled the mod for editor levels since it wasn't really usable (this also changes the location of the other save files)
+- Now when you delete a level the checkpoint saves of that level are also deleted (configurable) \[[Issue #25](https://github.com/Kevadroz/PracticeCheckpointPermanence/issues/25)\]
 - Added handling for invalid save files
 - Fixed the buttons in the Checkpoint Manager list being clickable outside the list
 - Make the active checkpoint be on top of the others on the progress bar